codeclimate number of rows
