+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_core/_events.py
@@ -1,6 +1,5 @@
 # pylint: disable=relative-beyond-top-level
 
-import json
 import logging
 from typing import Any, Final, Optional, cast
 from uuid import uuid4
@@ -20,7 +19,6 @@ from models_library.services import ServiceKeyVersion
 from pydantic import PositiveFloat
 from servicelib.fastapi.long_running_tasks.client import TaskId
 from servicelib.json_serialization import json_dumps
-from servicelib.utils import logged_gather
 from simcore_postgres_database.models.comp_tasks import NodeClass
 from simcore_service_director_v2.utils.dict_utils import nested_update
 from tenacity._asyncio import AsyncRetrying
@@ -28,8 +26,7 @@ from tenacity.before_sleep import before_sleep_log
 from tenacity.stop import stop_after_delay
 from tenacity.wait import wait_fixed
 
-from .....core.errors import NodeRightsAcquireError
-from .....core.settings import AppSettings, DynamicSidecarSettings
+from .....core.settings import DynamicSidecarSettings
 from .....models.schemas.dynamic_services import DynamicSidecarStatus, SchedulerData
 from .....models.schemas.dynamic_services.scheduler import (
     DockerContainerInspect,
@@ -40,7 +37,6 @@ from ....catalog import CatalogClient
 from ....db.repositories.internet_to_groups import InternetToGroupsRepository
 from ....db.repositories.projects import ProjectsRepository
 from ....director_v0 import DirectorV0Client
-from ....node_rights import NodeRightsManager
 from ....rabbitmq import RabbitMQClient
 from ...api_client import (
     BaseClientHTTPError,
@@ -65,12 +61,12 @@ from ...docker_service_specs import (
 from ...errors import EntrypointContainerNotFoundError, UnexpectedContainerStatusError
 from ._abc import DynamicSchedulerEvent
 from ._events_utils import (
-    RESOURCE_STATE_AND_INPUTS,
     are_all_user_services_containers_running,
     attach_project_networks,
     attempt_pod_removal_and_data_saving,
     get_director_v0_client,
     parse_containers_inspect,
+    prepare_services_environment,
     wait_for_sidecar_api,
 )
 
@@ -378,71 +374,7 @@ class PrepareServicesEnvironment(DynamicSchedulerEvent):
 
     @classmethod
     async def action(cls, app: FastAPI, scheduler_data: SchedulerData) -> None:
-        app_settings: AppSettings = app.state.settings
-        dynamic_sidecar_client = get_dynamic_sidecar_client(app)
-        dynamic_sidecar_endpoint = scheduler_data.endpoint
-        dynamic_sidecar_settings: DynamicSidecarSettings = (
-            app_settings.DYNAMIC_SERVICES.DYNAMIC_SIDECAR
-        )
-
-        async def _pull_outputs_and_state():
-            tasks = [
-                dynamic_sidecar_client.pull_service_output_ports(
-                    dynamic_sidecar_endpoint
-                )
-            ]
-            # When enabled no longer downloads state via nodeports
-            # S3 is used to store state paths
-            if not app_settings.DIRECTOR_V2_DEV_FEATURE_R_CLONE_MOUNTS_ENABLED:
-                tasks.append(
-                    dynamic_sidecar_client.restore_service_state(
-                        dynamic_sidecar_endpoint
-                    )
-                )
-
-            await logged_gather(*tasks, max_concurrency=2)
-
-            # inside this directory create the missing dirs, fetch those form the labels
-            director_v0_client: DirectorV0Client = get_director_v0_client(app)
-            simcore_service_labels: SimcoreServiceLabels = (
-                await director_v0_client.get_service_labels(
-                    service=ServiceKeyVersion(
-                        key=scheduler_data.key, version=scheduler_data.version
-                    )
-                )
-            )
-            service_outputs_labels = json.loads(
-                simcore_service_labels.dict().get("io.simcore.outputs", "{}")
-            ).get("outputs", {})
-            logger.debug(
-                "Creating dirs from service outputs labels: %s",
-                service_outputs_labels,
-            )
-            await dynamic_sidecar_client.service_outputs_create_dirs(
-                dynamic_sidecar_endpoint, service_outputs_labels
-            )
-
-            scheduler_data.dynamic_sidecar.is_service_environment_ready = True
-
-        if dynamic_sidecar_settings.DYNAMIC_SIDECAR_DOCKER_NODE_RESOURCE_LIMITS_ENABLED:
-            node_rights_manager = NodeRightsManager.instance(app)
-            assert scheduler_data.dynamic_sidecar.docker_node_id  # nosec
-            try:
-                async with node_rights_manager.acquire(
-                    scheduler_data.dynamic_sidecar.docker_node_id,
-                    resource_name=RESOURCE_STATE_AND_INPUTS,
-                ):
-                    await _pull_outputs_and_state()
-            except NodeRightsAcquireError:
-                # Next observation cycle, the service will try again
-                logger.debug(
-                    "Skip saving service state for %s. Docker node %s is busy. Will try later.",
-                    scheduler_data.node_uuid,
-                    scheduler_data.dynamic_sidecar.docker_node_id,
-                )
-                return
-        else:
-            await _pull_outputs_and_state()
+        await prepare_services_environment(app, scheduler_data)
 
 
 class CreateUserServices(DynamicSchedulerEvent):
